Test262AgentSleep should not assume that its argument is int.
(cherry picked from commit 47aab1c0458303a7bb22fc6de283beda2a470809)

+++ graal-js/src/com.oracle.truffle.js/src/com/oracle/truffle/js/runtime/util/DebugJSAgent.java
@@ -191,7 +191,7 @@ public class DebugJSAgent extends JSAgent {
     }
 
     @TruffleBoundary
-    public void sleep(int time) {
+    public void sleep(long time) {
         try {
             Thread.sleep(time);
         } catch (InterruptedException e) {
